+++ packages/client/src/v2-events/hooks/useDefaultValues.ts
@@ -28,9 +28,9 @@ export function handleDefaultValue({
 }) {
   const defaultValue = field.defaultValue
 
-  if (isFieldConfigDefaultValue(field.defaultValue)) {
+  if (isFieldConfigDefaultValue(defaultValue)) {
     return replacePlaceholders({
-      fieldType: field.type,
+      field,
       defaultValue,
       systemVariables
     })